Phonebook updated

+++ part2/phonebook/src/components/SavedNumbers.jsx
@@ -1,5 +1,4 @@
 /* eslint-disable no-unused-vars */
-import { useEffect } from 'react';
 import serverDB from '../services/communication'
 
 const SavedNumbers = ({people, setPersons}) => {
@@ -8,12 +7,13 @@ const SavedNumbers = ({people, setPersons}) => {
     <>
         {people.map((contact, id) => 
         <div key={contact.name + id + 'container'} style={{margin: '.75em 0'}}>
-            <p key={contact.name + id} style={{display: "inline"}}>{contact.name} {contact.number}&nbsp;&nbsp;&nbsp;&nbsp;</p> 
+            <p key={contact.name + id} style={{display: "inline"}}>{contact.name}&nbsp;&nbsp;{contact.number}&nbsp;&nbsp;&nbsp;&nbsp;</p> 
             <button key={contact.name + id + 'deleteButton'} onClick={() => {
-                let updatedPeople = [...people];
-                updatedPeople.splice(id, 1)
-                serverDB.deleteContact(contact.id).then(() => {setPersons(updatedPeople)})
-                setPersons(updatedPeople)
+                if (confirm(`Delete ${contact.name}?`)){
+                    let updatedPeople = [...people];
+                    updatedPeople.splice(id, 1)
+                    serverDB.deleteContact(contact.id).then(() => {setPersons(updatedPeople)})
+                }
             }}>delete</button>
         </div>
         )}
